fixata edit

+++ app/Http/Controllers/GuestHouseController.php
@@ -11,22 +11,22 @@ class GuestHouseController extends Controller{
     public function __construct(){
         $this->middleware('auth')->except('index', 'show');
     }
-
+    
     public function create() {
         return view('product.create');
     }
-
+    
     /**
-     * Store a newly created resource in storage.
-     */
+    * Store a newly created resource in storage.
+    */
     public function store(Request $request){
         
         
     }
-
+    
     /**
-     * Display the specified resource.
-     */
+    * Display the specified resource.
+    */
     public function show($id){
         $house = GuestHouse::findOrFail($id);
         //utente proprietario vede i propri non accettati
@@ -54,32 +54,46 @@ class GuestHouseController extends Controller{
             //dd($guest_houses);
             return redirect(route('home'))->with('messageNotFound', 'Articolo non trovato')->with(compact('guest_houses'));
         }
-
+        
     }
-
+    
     public function edit(GuestHouse $house){
         $locations = Location::all();
+        
+        if(Auth::user() && Auth::user()->name==$house->user->name)
+        
         return view('product.edit', compact('house','locations'));
+        $guest_houses= GuestHouse::all();
+        return redirect(route('home'))->with(compact('guest_houses'));
     }
-
+    
     /**
-     * Update the specified resource in storage.
-     */
+    * Update the specified resource in storage.
+    */
     public function update(Request $request, GuestHouse $guestHouse)
     {
         //
     }
-
+    
     /**
-     * Remove the specified resource from storage.
-     */
+    * Remove the specified resource from storage.
+    */
     public function destroy(GuestHouse $house)
     {
-        $house->delete();
-        return redirect(route('index'))->with('message', 'prodotto eliminato correttamente');
+        // $locations = Location::all();
+        
+        if(Auth::user() && Auth::user()->name==$house->user->name){
+        
+          $house->delete();
+          return redirect(route('index'))->with('message', 'prodotto eliminato correttamente');
+        }
+        // $guest_houses= GuestHouse::all();
+        // return redirect(route('home'))->with(compact('guest_houses'));
+
+       
     }
-
-
-
+    
+    
+    
     
 }
